Move protocol infrastructure into its own submodule
